iommu/vt-d: Retire struct intel_svm
The sva domain is a kind of iommu_domain which allows the processor and
device to use the same virtual addresses that are managed by system
memory management. As a result, use dmar_domain to stand for a sva domain
instead of using struct intel_svm to avoid duplicating code.

Besides, allocate sva domain in domain_alloc_sva() callback which allows
the memory management notifier lifetime to follow the lifetime of the
iommu_domain.

Signed-off-by: Tina Zhang <tina.zhang@intel.com>