iommu/vt-d: Add domain_type_is_sva helper function
Introduce a helper function to tell if the type of a dmar_domain is sva
or not. In general, comparing with normal iommu_domain, a sva domain
shouldn't be special other than having a different source for the page
table. Thus, this helper function should be used as little as possible
(i.e. only in the cases related to handling page table source).

Signed-off-by: Tina Zhang <tina.zhang@intel.com>